fix build error
