+++ src/utils/Post.ts
@@ -3,11 +3,8 @@ import fs from 'fs'
 
 import memoize from 'memoizee'
 import frontMatter from 'front-matter'
-import { GetStaticPaths, GetStaticProps } from 'next'
-import { ParsedUrlQuery } from 'node:querystring'
 
 import { FrontMatter, Post, TagWithCount } from '../common/types'
-import { parseMarkdownToMDX } from '../utils/Markdown'
 
 const DIR_REPLACE_STRING = '/posts'
 
@@ -119,51 +116,4 @@ export async function retreiveAllPosts(): Promise<Array<Post>> {
   }
 
   return posts.sort((a, b) => b.frontMatter.date - a.frontMatter.date)
-}
-
-export const getPostStaticPaths = (
-  year: string,
-): GetStaticPaths => async () => {
-  const posts = (await getAllPosts()).filter(({ fields: { slug } }) =>
-    slug.startsWith(year),
-  )
-
-  const paths: Array<{
-    params: { slugs: string[] }
-  }> = posts.reduce<Array<{ params: { slugs: string[] } }>>(
-    (prev, { fields: { slug } }) => {
-      const [...slugs] = `${slug.replace('.md', '')}`.split('/')
-
-      prev.push({ params: { slugs } })
-      return prev
-    },
-    [],
-  )
-
-  return {
-    paths,
-    fallback: 'blocking',
-  }
-}
-
-interface SlugInterface extends ParsedUrlQuery {
-  [key: string]: string | string[] | undefined
-  slugs: string[]
-}
-
-export const getPostStaticProps = (year: string): GetStaticProps => async ({
-  params,
-}) => {
-  const { slugs } = params as SlugInterface
-
-  const slug = [year, ...(slugs as string[])].join('/')
-  const posts = await getAllPosts()
-  const post = posts.find(({ fields: { slug: postSlug } }) => postSlug === slug)
-
-  return {
-    props: {
-      post: post,
-      mdx: post && (await parseMarkdownToMDX(post.body, post.path)),
-    },
-  }
 }
